Created the init and find API calls in the mapactivity.

+++ app/src/main/java/com/example/pickup/MapPresenter.java
@@ -18,7 +18,7 @@ public class MapPresenter {
     // Called by server
     // Create event object
     // Create marker
-    public void makeEvent(float latitude, float longitude, int minPeople, int maxPeople, String description) {
+    public void makeEvent(double latitude, double longitude, int minPeople, int maxPeople, String description) {
         Event myEvent = new Event(currID++, latitude, longitude, minPeople, maxPeople, description);
         events.add(myEvent);
         Marker marker = myMap.drawPickUpMarker(myEvent); //Unsure if needed yet
